+++ packages/vchart-types/types/index-harmony.d.ts
@@ -1,0 +1,8 @@
+import { VChart } from './core';
+import { DefaultTicker } from '@visactor/vrender-core';
+export { VChart, DefaultTicker };
+export default VChart;
+export * from './core';
+export * from './plugin/chart';
+export * from './plugin/components/tooltip-handler';
+export * from './plugin/components/axis-sync';