+++ packages/vchart/src/plugin/components/tooltip-handler/utils/style.ts
@@ -1,4 +1,4 @@
-import { isArray, isValid, isValidNumber, lowerCamelCaseToMiddle, normalizePadding } from '@visactor/vutils';
+import { isArray, isValid, isValidNumber, normalizePadding } from '@visactor/vutils';
 import type { ITooltipSpec, ITooltipTextTheme, ITooltipTheme } from '../../../../component/tooltip';
 const DEFAULT_SHAPE_SPACING = 8;
 const DEFAULT_KEY_SPACING = 26;
@@ -15,16 +15,35 @@ export const getPixelPropertyStr = (num?: number | number[], defaultStr?: string
 };
 
 export const getTextStyle = (style: ITooltipTextTheme = {}) => {
-  const textStyle: Partial<CSSStyleDeclaration> = {
-    color: style.fill ?? style.fontColor,
-    fontFamily: style.fontFamily,
-    fontSize: getPixelPropertyStr(style.fontSize as number),
-    fontWeight: style.fontWeight as string,
-    textAlign: style.textAlign,
-    maxWidth: getPixelPropertyStr(style.maxWidth),
-    whiteSpace: style.multiLine ? 'initial' : 'nowrap',
-    wordBreak: style.multiLine ? style.wordBreak ?? 'break-word' : 'normal'
-  };
+  const textStyle: Partial<CSSStyleDeclaration> = {};
+
+  if (isValid(style.fontFamily)) {
+    textStyle.fontFamily = style.fontFamily;
+  }
+  const color = style.fill ?? style.fontColor;
+
+  if (isValid(color)) {
+    textStyle.color = color;
+  }
+  if (isValid(style.fontWeight)) {
+    textStyle.fontWeight = style.fontWeight as string;
+  }
+  if (isValid(style.textAlign)) {
+    textStyle.textAlign = style.textAlign as string;
+  }
+  if (isValid(style.fontSize)) {
+    textStyle.fontSize = getPixelPropertyStr(style.fontSize as number);
+  }
+  if (isValid(style.maxWidth)) {
+    textStyle.maxWidth = getPixelPropertyStr(style.maxWidth as number);
+  }
+  if (style.multiLine) {
+    textStyle.whiteSpace = 'initial';
+    textStyle.wordBreak = style.wordBreak ?? 'break-word';
+  } else {
+    textStyle.wordBreak = 'normal';
+    textStyle.whiteSpace = 'nowrap';
+  }
 
   return textStyle;
 };
@@ -130,17 +149,4 @@ export function setStyleToDom(dom: HTMLElement, style: Partial<CSSStyleDeclarati
   Object.keys(style).forEach(key => {
     (dom.style as any)[key] = (style as any)[key];
   });
-}
-
-export function cssToStyleString(style: Partial<CSSStyleDeclaration>) {
-  let str = '';
-
-  style &&
-    Object.keys(style).forEach(k => {
-      if (isValid((style as any)[k])) {
-        str += `${lowerCamelCaseToMiddle(k)}:${(style as any)[k]};`;
-      }
-    });
-
-  return str;
 }